Change array join to safe_join
join concatenates array elements to create a string with the optional
specified separator. Unfortunately, the resulting string is assumed not
HTML-safe, and so Rails escapes it to be safe. Rails offers a better
alternative, safe_join, that will concatenate elements while retaining
each element’s HTML-safety and returning an HTML-safe string.