Allow arrays inside nested parameters
When one scope is handled by multiple fields, we need to add each of
those fields to the permitted params. However, when one of those inner
fields includes multiple-select fields, we don't instruct Strong
Parameters that those params expect arrays. As a result, those keys are
dropped.

This commit fixes that, recursively handling each nested param, so that
arrays and other nested fields are supported.